+++ cpp/include/cudf/utilities/traits.hpp
@@ -17,6 +17,7 @@
 #pragma once
 
 #include <cudf/types.hpp>
+#include <cudf/wrappers/timestamps.hpp>
 #include <cudf/utilities/type_dispatcher.hpp>
 #include <cudf/strings/string_view.cuh>
 
@@ -82,7 +83,7 @@ struct is_numeric_impl {
  *
  * "Numeric" types are fundamental integral/floating point types such as `INT*`
  * or `FLOAT*`. Types that wrap a numeric type are not considered numeric, e.g.,
- *`TIMESTAMP` or `DATE32`.
+ *`TIMESTAMP`.
  *
  * @param type The `data_type` to verify
  * @return true `type` is numeric
@@ -90,6 +91,43 @@ struct is_numeric_impl {
  *---------------------------------------------------------------------------**/
 constexpr inline bool is_numeric(data_type type) {
   return cudf::experimental::type_dispatcher(type, is_numeric_impl{});
+}
+
+/**---------------------------------------------------------------------------*
+ * @brief Indicates whether the type `T` is a timestamp type.
+ *
+ * @tparam T  The type to verify
+ * @return true `T` is a timestamp
+ * @return false  `T` is not a timestamp
+ *---------------------------------------------------------------------------**/
+template <typename T>
+constexpr inline bool is_timestamp() {
+  return (
+    std::is_same<cudf::timestamp_D, T>::value ||
+    std::is_same<cudf::timestamp_s, T>::value ||
+    std::is_same<cudf::timestamp_ms, T>::value ||
+    std::is_same<cudf::timestamp_us, T>::value ||
+    std::is_same<cudf::timestamp_ns, T>::value );
+}
+
+struct is_timestamp_impl {
+  template <typename T>
+  bool operator()() {
+    return is_timestamp<T>();
+  }
+};
+
+/**---------------------------------------------------------------------------*
+ * @brief Indicates whether `type` is a timestamp `data_type`.
+ *
+ * "Timestamp" types are int32_t or int64_t durations since the unix epoch.
+ *
+ * @param type The `data_type` to verify
+ * @return true `type` is a timestamp
+ * @return false `type` is not a timestamp
+ *---------------------------------------------------------------------------**/
+constexpr inline bool is_timestamp(data_type type) {
+  return cudf::exp::type_dispatcher(type, is_timestamp_impl{});
 }
 
 /**---------------------------------------------------------------------------*
@@ -105,7 +143,7 @@ template <typename T>
 constexpr inline bool is_fixed_width() {
   // TODO Add fixed width wrapper types
   // Is a category fixed width?
-  return cudf::is_numeric<T>();
+  return cudf::is_numeric<T>() || cudf::is_timestamp<T>();
 }
 
 struct is_fixed_width_impl {